#38 Installation module is refactored
