+++ Common/configuration/orderer.yaml
@@ -91,7 +91,7 @@ General:
     # initializing the orderer system channel and BootstrapMethod is set to
     # "file".  The bootstrap file can be the genesis block, and it can also be
     # a config block for late bootstrap of some consensus methods like Raft.
-    # Generate a genesis block by updating $FABRIC_CFG_PATH/configtx.yaml and
+    # Generate a genesis block by updating $FABRIC_CFG_PATH/local-configtx.yaml and
     # using configtxgen command with "-outputBlock" option.
     # Defaults to file "genesisblock" (in $FABRIC_CFG_PATH directory) if not specified.
     BootstrapFile:
