+++ Common/index.js
@@ -1,5 +1,6 @@
 
-module.exports.Installation = require("./lib/installation");
+module.exports.Installation = require("./lib/installation").Installation;
 module.exports.CaNode = require("./lib/CaNode");
 module.exports.PeerNode = require("./lib/PeerNode");
 module.exports.OrdererNode = require(`./lib/OrdererNode`);
+module.exports.DockerApi = require(`./lib/dockerApi`);
