+++ Common/lib/installation/index.js
@@ -1,0 +1,1 @@
+module.exports.Installation = require(`./Installation`);
