+++ local.peer1.org2.env
@@ -1,9 +1,9 @@
 export CORE_PEER_LOCALMSPID=Org2
 export CORE_PEER_ID=cli
 export CORE_PEER_ADDRESS=0.0.0.0:7058
-export FABRIC_CFG_PATH=/home/test/work/fabric-config
-export ORDERER_TLS_CA=/home/test/ttz/orderer-tls-ca-cert.pem
-export CORE_PEER_TLS_ROOTCERT_FILE=/home/test/ttz/Org2/msp/tlscacerts/tls-ca-cert.pem
-export CORE_PEER_MSPCONFIGPATH=/home/test/ttz/Org2/fabric-ca/client/org-ca/org-admin/msp/
+export FABRIC_CFG_PATH=/home/anil/config
+export ORDERER_TLS_CA=/home/anil/ttz/orderer-tls-ca-cert.pem
+export CORE_PEER_TLS_ROOTCERT_FILE=/home/anil/ttz/Org2/msp/tlscacerts/tls-ca-cert.pem
+export CORE_PEER_MSPCONFIGPATH=/home/anil/ttz/Org2/fabric-ca/client/org-ca/org-admin/msp/
 export CORE_PEER_TLS_ENABLED=true
 export ORDERER_ADDRESS=0.0.0.0:7050
